requirements.txt
flask
ultralytics
werkzeug
pillow
flask-cors